pyproject.toml
[tool.poetry]
name = "mar-2023-pyton-drf444"
version = "0.1.0"
description = ""
authors = ["Vitaliy Dechyshyn <demchyshyn.v87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
mysqlclient = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"


[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
